Naming + minor renames (#47)
- `ui::search::setup_search_entry` → `ui::widgets::build_search_entry`.
  The existing factories in `widgets.rs` already use the `build_*`
  prefix (`app_icon_button` aside, but that one is more descriptive
  than a generic `build`), and the search-entry factory is a one-line
  widget builder that fits cleanly next to them.
- `clear_drawer_state` (3-line wrapper around a single
  `search_entry.set_text("")`) inlined at its lone call site in
  `listeners/commands.rs`. The function name oversold what it did.
- Folded `ui/search.rs` into `ui/app_grid.rs`. The module had two
  unrelated functions (`subsequence_match` + `setup_search_entry`)
  and shared a name prefix with `ui/search_handler.rs`, which made
  "which one do I use?" the wrong question to have to ask.
  `subsequence_match` lives next to its sole caller (the search-mode
  filter in `build_app_flow_box`) and the search-entry factory moved
  to `widgets.rs` per the rename above.
